+added MedianFilterSquare3x3();

+++ trunk/src/Simd/SimdResizeBilinear.h
@@ -32,7 +32,7 @@ namespace Simd
     {
         void ResizeBilinear(
             const uchar *src, size_t srcWidth, size_t srcHeight, size_t srcStride,
-            uchar *dst, size_t dstWidth, size_t dstHeight, size_t dstStride, int channelCount);
+            uchar *dst, size_t dstWidth, size_t dstHeight, size_t dstStride, size_t channelCount);
     }
 
 #ifdef SIMD_SSE2_ENABLE
@@ -46,6 +46,6 @@ namespace Simd
 
 	void ResizeBilinear(
 		const uchar *src, size_t srcWidth, size_t srcHeight, size_t srcStride,
-		uchar *dst, size_t dstWidth, size_t dstHeight, size_t dstStride, int channelCount);
+		uchar *dst, size_t dstWidth, size_t dstHeight, size_t dstStride, size_t channelCount);
 }
 #endif//__SimdResizeBilinear_h__
